requirements.txt
Flask==2.0.1
Flask-SocketIO==5.1.1
Flask-CORS==4.0.0
python-socketio==5.4.0
opencv-python-headless==4.5.3.56
numpy==1.21.2
Pillow==8.3.2
tensorflow-cpu==2.7.0
gunicorn==20.1.0
eventlet==0.33.0